+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.6.2
+
+* **Support opinionated writes in databricks delta table connector**
+* **Update databricks volume connector to emit user agent**
+* **Delete previous content from databricks delta tables**
+
 ## 0.6.1
 
 ### Fixes